adds a separate "genpkg" executable
genpkg takes a similar "-o" option, which in this case specifies an
output *directory*, then a package name and replacement types.

It calls out to "go get" to ensure the package is fetched, locates the
package according to $GOPATH, then effectively runs all the .go files
through gengen, dumping the output files to the output dir (which
defauls to the current directory).

The purpose is to enable type-agnostic datastructures and algorithms to
not just be *written*, but also *published*. Taking the btree package
from the examples/ directory, to get a subpackage of btree<string,int>
I could just put this in my project:

    //go:generate genpkg -o ./btree github.com/joeshaw/gengen/examples/btree string int

Then "go generate" would fetch that third-party generic package and
codegen the type-specific package I need for my project from it.

To make this work I had to separate out the core functionality into a
"genlib" package so it could be imported into both utilities.